+++ Makefile
@@ -13,7 +13,8 @@ VPATH = src bonus src/core src/init src/movement src/parsing \
 	src/parsing/create_and_validate_map src/parsing/parse_info src/draw
 
 TEST_NAME = tester
-TEST_SRC = tests/tester.c tests/exec_program.c
+TEST_SRC = tests/tests.c tests/exec_program.c \
+		tests/check_invalid.c tests/run_cub3d.c
 TEST_OBJ = $(TEST_SRC:%.c=$(OBJ_DIR)/%.o)
 TEST_DEP = $(TEST_SRC:%.c=$(OBJ_DIR)/%.d)
 TEST_CFLAGS = -MMD
